pyproject.toml
[project]
name = "agentic-design-patterns"
version = "0.1.0"
description = "Agentic Design Patterns"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "groq>=0.9.0",
    "jupyter >=1.0.0",
    "python-dotenv >=1.0.1",
    "graphviz >=0.20.3",
    "pytest>=8.3.4",
    "ruff >=0.7.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.7.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/utils"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
